+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Fix setting next fan profile
+
 ## [v5.0.0]
 ### Added
 - Gnome 45 plugin